real

+++ CUE4Parse/UE4/Assets/Objects/FCompressedChunk.cs
@@ -5,8 +5,8 @@ namespace CUE4Parse.UE4.Assets.Objects
 {
     public readonly struct FCompressedChunk
     {
-        public readonly long UncompressedOffset;
-        public readonly long UncompressedSize;
+        public readonly int UncompressedOffset;
+        public readonly int UncompressedSize;
         public readonly int CompressedOffset;
         public readonly int CompressedSize;
 
@@ -14,8 +14,8 @@ namespace CUE4Parse.UE4.Assets.Objects
         {
             if (Ar.Game == EGame.GAME_RocketLeague)
             {
-                UncompressedOffset = Ar.Read<long>();
-                UncompressedSize = Ar.Read<long>();
+                UncompressedOffset = (int)Ar.Read<long>();
+                UncompressedSize = (int)Ar.Read<long>();
                 goto SkipToCompressed;
             }
